#81: Drop test.monitors; apply removestar
