+++ test/test_monge_ampere.py
@@ -2,10 +2,35 @@ import unittest
 from unittest.mock import MagicMock
 
 import numpy as np
-from monitors import *
 from parameterized import parameterized
+import pytest
+from firedrake import (
+    Constant,
+    DirichletBC,
+    File,
+    SpatialCoordinate,
+    UnitCubeMesh,
+    UnitSquareMesh,
+    assemble,
+    ds,
+)
+from ufl import as_vector, cosh, dot
 
-from movement import *
+from movement import MongeAmpereMover
+
+
+def const_monitor(mesh):
+    return Constant(1.0)
+
+
+def ring_monitor(mesh):
+    alpha = Constant(10.0)  # amplitude
+    beta = Constant(200.0)  # width
+    gamma = Constant(0.15)  # radius
+    dim = mesh.geometric_dimension()
+    xyz = SpatialCoordinate(mesh) - as_vector([0.5] * dim)
+    r = dot(xyz, xyz)
+    return Constant(1.0) + alpha / cosh(beta * (r - gamma)) ** 2
 
 
 class TestMongeAmpere(unittest.TestCase):
